logging: accept a unit suffix in virtlogd's max_size
Switch max_size from virConfGetValueSizeT() to virConfGetValueBytes(),
so it accepts "2MiB" the same way domain XML memory sizes do, instead
of only a raw byte count.

virtlogd.aug accepts the quoted form too; its quoted branch requires a
unit suffix, as one matching a bare number would overlap the integer
branch.

+++ src/logging/log_daemon_config.h
@@ -32,7 +32,7 @@ struct _virLogDaemonConfig {
     unsigned int admin_max_clients;
 
     size_t max_backups;
-    size_t max_size;
+    unsigned long long max_size;
 
     char *log_root;
     size_t max_age_days;
